[CIAS30-3631] Rename saga
