+++ app/containers/InterventionDetailsPage/saga.js
@@ -8,7 +8,7 @@ import {
   externalCopySessionSaga,
   sendInterventionInvitationsSaga,
   resendInterventionInvitationSaga,
-  fetchInterventionInvitesSaga,
+  fetchInterventionInvitations,
   generateConversationsTranscriptSaga,
   changeInterventionNarratorSaga,
   editShortLinksSaga,
@@ -28,7 +28,7 @@ export default function* interventionDetailsPageSagas() {
     sendInterventionInvitationsSaga(),
     resendInterventionInvitationSaga(),
     externalCopySessionSaga(),
-    fetchInterventionInvitesSaga(),
+    fetchInterventionInvitations(),
     generateConversationsTranscriptSaga(),
     changeInterventionNarratorSaga(),
     editShortLinksSaga(),
